Fix ROUND type error in compare endpoint (rooms::float -> rooms::numeric)
PostgreSQL ROUND(double precision, int) doesn't exist. Using numeric
type which ROUND supports with precision argument.

+++ api.py
@@ -968,7 +968,7 @@ def compare_neighborhoods():
                     PERCENTILE_CONT(0.5) WITHIN GROUP (ORDER BY price_numeric)::int as median_price,
                     ROUND(AVG(CASE WHEN size_sqm ~ '^[0-9]+$' AND size_sqm::int > 0
                         THEN price_numeric::float / size_sqm::int END))::int as avg_price_per_sqm,
-                    ROUND(AVG(CASE WHEN rooms ~ '^[0-9.]+$' THEN rooms::float END), 1) as avg_rooms,
+                    ROUND(AVG(CASE WHEN rooms ~ '^[0-9.]+$' THEN rooms::numeric END), 1) as avg_rooms,
                     ROUND(AVG(CASE WHEN size_sqm ~ '^[0-9]+$' THEN size_sqm::int END))::int as avg_sqm,
                     COUNT(*) FILTER (WHERE is_merchant) as agents,
                     COUNT(*) FILTER (WHERE NOT is_merchant) as private
